Fix Issue #11850 by extending move-model AST visitor to be able to hook BeforeBody in a Block expression (#11916)
Fix Issue #11850, a bug in flow_insensitive_checkers struct SymbolVisitor which
should process the binding of a Block expression (which is the RHS
of a let expression's variable assignment) in the outer context,
while processing the body in the context with the let variable(s)
bound. This was achieved by extending AST's ExpData::visit_pre_post
to visit_positions which provides hooks not just Pre and Post
subexpression traversal, but also at useful control points: PreBody
for Block expressions, and a few other points which will be useful
in other analyses.