[compiler v2] New Borrow Analysis (#11892)
* [compiler v2] New Borrow Analysis

This is a major rewrite of the reference safety analysis. This PR closes #11251 which shows the problem with the previous version: essentially, it is necessary to allow for a sequence of instructions an "unsafe" borrow graph and only enforce safety at particular instruction boundaries. The simplest example to demonstrate the problem is `f(&mut s.x, &mut s.y)`. If broken down to instructions, at some point there will be a reference to `&mut s` where at the same time there is another one `&mut s.f`, which is unsafe if those refs would be passed as function parameters. The rewrite does the following:

- Instead of checking safety of borrow operations at the time they are performed, the graph is now constructed first, and later checked for safety at particular instructions (read ref, write ref, and function calls)
- The borrow graph stays mostly the same, instead of Skip edges for graph glueing which became overly complex, glueing is now performed by graph renaming.
- There are various enhancements in error reporting (and potentially because of the delayed borrow checking also some regressions)

It might be useful to review the new version without diff from scratch.

* Adding some ASCII graphics to the documentation of the analysis domain.

* Adding a copy of the new analysis in `reference_safety_processor_v2.rs`. This allows for review from the basics.

* Updating reference safety:

- Unifying the three borrow safety checks now in `check_graph_safety`. Documenting them more concisely in the module overview with a sketch of a more formal semantics.
- Reducing graph safety checks to argument locals, removing some false positives.
- Covering a missed violation of using the same reference multiple times in an argument list

* Minor renamings, fixing tests

* Addressing reviewer comments

* Updating reference_safety_processor_v2, renaming more local to temp

* Fixing tests

* Change permutations to combinations

* Addressing reviewer comments.

* Final touches for merge