[move-2] Allow variables and functions be called `match` in Move-2 (#14049)
* [move-2] Allow variables and functions be called `match` in Move-2

Fixes #14038.

The introduction of the `match (exp) { .. }` expression accidentally disabled the `match` name to be used for functions or variables. This PR adds context sensitivity to the parser to distinguish, for example, `match(x, y)` (a function call) from `match (e) { .. }`. As we are using a LL(2) parser, this is achieved by factoring out the common prefix. There is no function call `match(e) {` possible, so we know it is a match expression once we see `{`.

* Adding a test for `fun match() {`